feat: add bookings and comments

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -7,11 +7,11 @@ import java.util.List;
 public interface UserService {
     List<UserDto> findAll();
 
-    UserDto findById(int userId);
+    UserDto findById(Long userId);
 
     UserDto create(UserDto userDto);
 
-    UserDto update(UserDto userDto, int userId);
+    UserDto update(UserDto userDto, Long userId);
 
-    void delete(int userId);
+    void delete(Long userId);
 }
